--- src/Enum/EnumSeat.hpp
@@ -1,11 +1,0 @@
-#ifndef ENUM_SEAT
-#define ENUM_SEAT
-
-enum class EnumSeat {
-    ECONOMICA,
-    ECONOMICA_PREMIUM,
-    EXECUTIVA,
-    PRIMEIRA_CLASSE
-};
-
-#endif
