+++ src/includes.hpp
@@ -13,8 +13,8 @@
 #include "Destination.hpp"
 #include "Client.hpp"
 #include "Pilot.hpp"
-#include "Enum/EnumFlight.hpp"
-#include "Enum/EnumSeat.hpp"
+#include "EnumFlight.hpp"
+#include "EnumSeat.hpp"
 
 
 #endif // INCLUDES_HPP
